Extract browser-only helper functions from shared utils file to their own files
