+++ lib/utils.js
@@ -92,67 +92,6 @@ exports.clean = function(str) {
 };
 
 /**
- * Parse the given `qs`.
- *
- * @private
- * @param {string} qs
- * @return {Object}
- */
-exports.parseQuery = function(qs) {
-  return qs
-    .replace('?', '')
-    .split('&')
-    .reduce(function(obj, pair) {
-      var i = pair.indexOf('=');
-      var key = pair.slice(0, i);
-      var val = pair.slice(++i);
-
-      // Due to how the URLSearchParams API treats spaces
-      obj[key] = decodeURIComponent(val.replace(/\+/g, '%20'));
-
-      return obj;
-    }, {});
-};
-
-/**
- * Highlight the given string of `js`.
- *
- * @private
- * @param {string} js
- * @return {string}
- */
-function highlight(js) {
-  return js
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/\/\/(.*)/gm, '<span class="comment">//$1</span>')
-    .replace(/('.*?')/gm, '<span class="string">$1</span>')
-    .replace(/(\d+\.\d+)/gm, '<span class="number">$1</span>')
-    .replace(/(\d+)/gm, '<span class="number">$1</span>')
-    .replace(
-      /\bnew[ \t]+(\w+)/gm,
-      '<span class="keyword">new</span> <span class="init">$1</span>'
-    )
-    .replace(
-      /\b(function|new|throw|return|var|if|else)\b/gm,
-      '<span class="keyword">$1</span>'
-    );
-}
-
-/**
- * Highlight the contents of tag `name`.
- *
- * @private
- * @param {string} name
- */
-exports.highlightTags = function(name) {
-  var code = document.getElementById('mocha').getElementsByTagName(name);
-  for (var i = 0, len = code.length; i < len; ++i) {
-    code[i].innerHTML = highlight(code[i].innerHTML);
-  }
-};
-
-/**
  * If a value could have properties, and has none, this function is called,
  * which returns a string representation of the empty value.
  *
