+++ test/unit/utils.spec.js
@@ -561,27 +561,6 @@ describe('lib/utils', function() {
     });
   });
 
-  describe('parseQuery()', function() {
-    var parseQuery = utils.parseQuery;
-    it('should get queryString and return key-value object', function() {
-      expect(parseQuery('?foo=1&bar=2&baz=3'), 'to equal', {
-        foo: '1',
-        bar: '2',
-        baz: '3'
-      });
-
-      expect(parseQuery('?r1=^@(?!.*\\)$)&r2=m{2}&r3=^co.*'), 'to equal', {
-        r1: '^@(?!.*\\)$)',
-        r2: 'm{2}',
-        r3: '^co.*'
-      });
-    });
-
-    it('should parse "+" as a space', function() {
-      expect(parseQuery('?grep=foo+bar'), 'to equal', {grep: 'foo bar'});
-    });
-  });
-
   describe('isPromise()', function() {
     it('should return true if the value is Promise-ish', function() {
       expect(
